requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
dill==0.2.7.1
Flask==0.12.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gevent==1.2.2
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpcio==1.7.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==3.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.5.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
